// RookieEcommerce.AdminSite/src/api/models/CreateProductImageCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateProductImageCommand
 */
export interface CreateProductImageCommand {
    /**
     * 
     * @type {string}
     * @memberof CreateProductImageCommand
     */
    productId?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductImageCommand
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateProductImageCommand
     */
    altText?: string | null;
}

/**
 * Check if a given object implements the CreateProductImageCommand interface.
 */
export function instanceOfCreateProductImageCommand(value: object): value is CreateProductImageCommand {
    return true;
}

export function CreateProductImageCommandFromJSON(json: any): CreateProductImageCommand {
    return CreateProductImageCommandFromJSONTyped(json, false);
}

export function CreateProductImageCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateProductImageCommand {
    if (json == null) {
        return json;
    }
    return {
        
        'productId': json['productId'] == null ? undefined : json['productId'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'altText': json['altText'] == null ? undefined : json['altText'],
    };
}

export function CreateProductImageCommandToJSON(json: any): CreateProductImageCommand {
    return CreateProductImageCommandToJSONTyped(json, false);
}

export function CreateProductImageCommandToJSONTyped(value?: CreateProductImageCommand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'productId': value['productId'],
        'imageUrl': value['imageUrl'],
        'altText': value['altText'],
    };
}