// RookieEcommerce.AdminSite/src/api/models/ProductImageDetailsDtoPaginationList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProductImageDetailsDto } from './ProductImageDetailsDto';
import {
    ProductImageDetailsDtoFromJSON,
    ProductImageDetailsDtoFromJSONTyped,
    ProductImageDetailsDtoToJSON,
    ProductImageDetailsDtoToJSONTyped,
} from './ProductImageDetailsDto';

/**
 * 
 * @export
 * @interface ProductImageDetailsDtoPaginationList
 */
export interface ProductImageDetailsDtoPaginationList {
    /**
     * 
     * @type {Array<ProductImageDetailsDto>}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    items?: Array<ProductImageDetailsDto> | null;
    /**
     * 
     * @type {number}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    pageNumber?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    totalCount?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    readonly totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    readonly hasPreviousPage?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProductImageDetailsDtoPaginationList
     */
    readonly hasNextPage?: boolean;
}

/**
 * Check if a given object implements the ProductImageDetailsDtoPaginationList interface.
 */
export function instanceOfProductImageDetailsDtoPaginationList(value: object): value is ProductImageDetailsDtoPaginationList {
    return true;
}

export function ProductImageDetailsDtoPaginationListFromJSON(json: any): ProductImageDetailsDtoPaginationList {
    return ProductImageDetailsDtoPaginationListFromJSONTyped(json, false);
}

export function ProductImageDetailsDtoPaginationListFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductImageDetailsDtoPaginationList {
    if (json == null) {
        return json;
    }
    return {
        
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(ProductImageDetailsDtoFromJSON)),
        'pageNumber': json['pageNumber'] == null ? undefined : json['pageNumber'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'totalCount': json['totalCount'] == null ? undefined : json['totalCount'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
        'hasPreviousPage': json['hasPreviousPage'] == null ? undefined : json['hasPreviousPage'],
        'hasNextPage': json['hasNextPage'] == null ? undefined : json['hasNextPage'],
    };
}

export function ProductImageDetailsDtoPaginationListToJSON(json: any): ProductImageDetailsDtoPaginationList {
    return ProductImageDetailsDtoPaginationListToJSONTyped(json, false);
}

export function ProductImageDetailsDtoPaginationListToJSONTyped(value?: Omit<ProductImageDetailsDtoPaginationList, 'totalPages'|'hasPreviousPage'|'hasNextPage'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(ProductImageDetailsDtoToJSON)),
        'pageNumber': value['pageNumber'],
        'pageSize': value['pageSize'],
        'totalCount': value['totalCount'],
    };
}